+++ EssentialFeed2Tests/Feed API/URLSessionHTTPClientTests.swift
@@ -152,16 +152,8 @@ class URLSessionHTTPClientTests: XCTestCase {
         return receivedResult
     }
     
-    private func anyURL() -> URL {
-        URL(string: "https://any-url.com")!
-    }
-    
     private func anyData() -> Data {
         Data("any data".utf8)
-    }
-    
-    private func anyNSError() -> NSError {
-        NSError(domain: "any error", code: 0, userInfo: nil)
     }
     
     private func nonHTTPURLResponse() -> URLResponse {
