+++ EssentialFeed2Tests/Feed Cache/LoadFeedFromCacheUseCaseTests.swift
@@ -186,45 +186,4 @@ class LoadFeedFromCacheUseCaseTests: XCTestCase {
         
         wait(for: [exp], timeout: 1)
     }
-    
-    private func anyNSError() -> NSError {
-        NSError(domain: "any error", code: 0, userInfo: nil)
-    }
-    
-    private func anyURL() -> URL {
-        URL(string: "https://any-url.com")!
-    }
-    
-    private func uniqueImage() -> FeedImage {
-        .init(
-            id: UUID(),
-            description: "any",
-            location: "any",
-            url: anyURL()
-        )
-    }
-    
-    private func uniqueImageFeed() -> (models: [FeedImage], local: [LocalFeedImage]) {
-        let models = [uniqueImage(), uniqueImage()]
-        let local = models.map {
-            LocalFeedImage(
-                id: $0.id,
-                description: $0.description,
-                location: $0.location,
-                url: $0.url
-            )
-        }
-        
-        return (models, local)
-    }
-}
-
-private extension Date {
-    func adding(days: Int) -> Date {
-        Calendar(identifier: .gregorian).date(byAdding: .day, value: days, to: self)!
-    }
-    
-    func adding(seconds: TimeInterval) -> Date {
-        self.addingTimeInterval(seconds)
-    }
 }
